Logger now fully implemented for printing to std out, bash script improved a bit.

+++ loggerdata/loggerdata.go
@@ -4,7 +4,7 @@ type LogMessage struct {
 	ClientName     string
 	DownloadedBits uint64
 	UploadedBits   uint64
-	TimeStamp      int
+	TimeStamp      float64
 }
 
 type LogReply struct {
